+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+19-2-2015
+=========
+* Added delete method to BagOfWords class
+* Refactored code
+* Added sub method to BafOfWords class
+
 18-2-2015
 =========
 * Initial version.